Update codec2.c
nse is unsigned int, document cast to float

+++ src/codec2.c
@@ -2285,7 +2285,7 @@ void codec2_load_codebook(struct CODEC2 *codec2_state, int num, char *filename) 
 
 float codec2_get_var(struct CODEC2 *codec2_state) {
     if (codec2_state->nse)
-        return codec2_state->se/codec2_state->nse;
+        return codec2_state->se / (float) codec2_state->nse;
     else
         return 0;
 }
